examples/simple02/requirements.txt
streamlit>=1.24.0
langchain>=0.0.300
langchain-openai>=0.0.2
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.5.0
